Move module to lib folder.
